backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
email-validator>=2.1.0
bcrypt>=4.1.0
PyJWT>=2.8.0
